+++ packages/node-core/src/utils/project.ts
@@ -232,7 +232,7 @@ export type IsRuntimeDs<DS> = (ds: DS) => ds is DS;
 // eslint-disable-next-line @typescript-eslint/require-await
 export async function insertBlockFiltersCronSchedules<DS extends BaseDataSource = BaseDataSource>(
   dataSources: DS[],
-  getBlockTimestamp: (height: number) => Promise<Date>,
+  getBlockTimestamp: (height: number) => Promise<Date | undefined>,
   isRuntimeDs: IsRuntimeDs<DS>,
   blockHandlerKind: string
 ): Promise<DS[]> {
@@ -247,8 +247,16 @@ export async function insertBlockFiltersCronSchedules<DS extends BaseDataSource 
             if (handler.kind === blockHandlerKind) {
               if (handler.filter?.timestamp) {
                 if (!timestampReference) {
-                  timestampReference = await getBlockTimestamp(startBlock);
+                  const blockTimestamp = await getBlockTimestamp(startBlock);
+                  if (!blockTimestamp) {
+                    throw new Error(
+                      `Could not apply cronSchedule, failed to get block timestamp for block ${startBlock}`
+                    );
+                  } else {
+                    timestampReference = blockTimestamp;
+                  }
                 }
+
                 let cronArr: number[][];
                 try {
                   cronArr = stringToArray(handler.filter.timestamp);